+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+- `String.quote/2` surrounds the given string with quotation marks.
+
 ## 4.0.0
 
 - Test with Elixir 1.19 and Erlang 28.1; drop support for Elixir 1.16 and Erlang 25.